refactor: convert babel-register to ESM
