+++ packages/babel-register/src/types.ts
