+++ packages/babel-register/src/worker/cache.ts
